Time: 31 ms (52.39%) | Memory: 44.3 MB (31.50%) - LeetSync

+++ BinarySearchTree/115-distinct-subsequences/distinct-subsequences.cpp
@@ -7,8 +7,8 @@ public:
         for(int i=0;i<=n;i++){
             dp[i][0]=1;
         }
-        for(long long i=1;i<=n;i++){
-            for(long long j=1;j<=m;j++){
+        for(int i=1;i<=n;i++){
+            for(int j=1;j<=m;j++){
                 if(s[i-1]==t[j-1]){
                     dp[i][j]=dp[i-1][j-1]+dp[i-1][j];
                 }else{
